+++ lincs/liblincs/learning/mrsort-by-weights-profiles-breed/breed/reinitialize-least-accurate.hpp
@@ -13,11 +13,13 @@ class ReinitializeLeastAccurate : public LearnMrsortByWeightsProfilesBreed::Bree
   explicit ReinitializeLeastAccurate(
     ModelsBeingLearned& models_being_learned_,
     LearnMrsortByWeightsProfilesBreed::ProfilesInitializationStrategy& profiles_initialization_strategy_,
+    LearnMrsortByWeightsProfilesBreed::WeightsOptimizationStrategy& weights_optimization_strategy_,
     unsigned count_
   ) :
     LearnMrsortByWeightsProfilesBreed::BreedingStrategy(true),
     models_being_learned(models_being_learned_),
     profiles_initialization_strategy(profiles_initialization_strategy_),
+    weights_optimization_strategy(weights_optimization_strategy_),
     count(count_)
   {}
 
@@ -27,6 +29,7 @@ class ReinitializeLeastAccurate : public LearnMrsortByWeightsProfilesBreed::Bree
  private:
   ModelsBeingLearned& models_being_learned;
   LearnMrsortByWeightsProfilesBreed::ProfilesInitializationStrategy& profiles_initialization_strategy;
+  LearnMrsortByWeightsProfilesBreed::WeightsOptimizationStrategy& weights_optimization_strategy;
   unsigned count;
 };