+++ doc-sources/changelog.rst
@@ -4,6 +4,12 @@
 Changelog
 =========
 
+Version 2.0.0a4 (2024-11-??)
+============================
+
+- Optimize WPB model weights before evaluating and returning them
+- **Breaking** Add parameter ``weights_optimization_strategy`` to the ``ReinitializeLeastAccurate`` breeding strategy (required to fix previous point)
+
 Version 2.0.0a3 (2024-10-10)
 ============================